requirements.txt
langchain==0.1.0
langchain-openai==0.0.2
python-whois==0.8.0
openai==1.12.0
llama-cpp-python==0.2.11
python-dotenv==1.0.0
gunicorn==20.1.0 